help: improvments
* inline beautify
* help improvments
* auto turn on setinline

+++ userbot/modules/__help.py
@@ -16,31 +16,51 @@
 # @Qulec tarafından yazılmıştır.
 # Thanks @Spechide.
 
-import logging
+from telethon.errors.rpcerrorlist import BotInlineDisabledError as noinline
+from telethon.errors.rpcerrorlist import YouBlockedUserError
 
 from userbot import BOT_USERNAME
-from userbot.events import register
-
-logging.basicConfig(
-    format="[%(levelname) 5s/%(asctime)s] %(name)s: %(message)s", level=logging.WARNING
-)
+from userbot import CMD_HANDLER as cmd
+from userbot import bot
+from userbot.events import man_cmd
 
 
-@register(outgoing=True, pattern=r"^\.helpme")
-async def yardim(event):
-    try:
-        tgbotusername = BOT_USERNAME
-        if tgbotusername is not None:
-            results = await event.client.inline_query(tgbotusername, "@UserButt")
+@bot.on(man_cmd(pattern="helpme", outgoing=True))
+async def _(event):
+    if event.fwd_from:
+        return
+    tgbotusername = BOT_USERNAME
+    chat = "@Botfather"
+    if tgbotusername is not None:
+        try:
+            results = await event.client.inline_query(tgbotusername, "@SharingUserbot")
             await results[0].click(
                 event.chat_id, reply_to=event.reply_to_msg_id, hide_via=True
             )
             await event.delete()
-        else:
-            await event.edit(
-                "**The bot doesn't work! Please set the Bot Token and Username correctly. The module has been stopped.**"
+        except noinline:
+            xx = await event.edit(
+                "**Inline Mode Tidak aktif.**\n__Sedang Menyalakannya, Harap Tunggu Sebentar...__"
             )
-    except Exception:
-        return await event.edit(
-            "**You cannot send inline results in this chat (caused by SendInlineBotResultRequest)**"
+            async with bot.conversation(chat) as conv:
+                try:
+                    first = await conv.send_message("/setinline")
+                    second = await conv.get_response()
+                    third = await conv.send_message(tgbotusername)
+                    fourth = await conv.get_response()
+                    fifth = await conv.send_message("Search")
+                    sixth = await conv.get_response()
+                    await bot.send_read_acknowledge(conv.chat_id)
+                except YouBlockedUserError:
+                    return await xx.edit("Unblock @Botfather first.")
+                await xx.edit(
+                    f"**Berhasil Menyalakan Mode Inline**\n\n**Ketik** `{cmd}helpme` **lagi untuk membuka menu bantuan.**"
+                )
+            await bot.delete_messages(
+                conv.chat_id,
+                [first.id, second.id, third.id, fourth.id, fifth.id, sixth.id],
+            )
+    else:
+        await event.edit(
+            "**ERROR:** Silahkan Cek Ulang `BOT_TOKEN` & `BOT_USERNAME` di Heroku."
         )
